+++ schemas/bedrock-account-http.js
@@ -1,7 +1,7 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import {schemas} from 'bedrock-validation';
+import {schemas} from '@bedrock/validation';
 
 export function create() {
   return {