+++ test/helpers.js
@@ -1,39 +1,37 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import * as bedrock from '@bedrock/core';
+import * as brAccount from '@bedrock/account';
+import * as database from '@bedrock/mongodb';
 
-const brAccount = require('bedrock-account');
-const database = require('bedrock-mongodb');
-const {util: {uuid}} = require('bedrock');
+const {util: {uuid}} = bedrock;
 
-const api = {};
-module.exports = api;
-
-api.createAccount = email => {
+export function createAccount(email) {
   const newAccount = {
     id: 'urn:uuid:' + uuid(),
     email
   };
   return newAccount;
-};
+}
 
 // called in test before hook
-api.prepareDatabase = async mockData => {
-  await api.removeCollections();
+export async function prepareDatabase(mockData) {
+  await removeCollections();
   await insertTestData(mockData);
-};
+}
 
 // called by prepareDatabase
-api.removeCollections = async (collectionNames = ['account']) => {
+export async function removeCollections(collectionNames = ['account']) {
   await database.openCollections(collectionNames);
   for(const collectionName of collectionNames) {
     await database.collections[collectionName].deleteMany({});
   }
-};
+}
 
-api.removeCollection =
-  async collectionName => api.removeCollections([collectionName]);
+export async function removeCollection(collectionName) {
+  return removeCollections([collectionName]);
+}
 
 // called by prepareDatabase
 async function insertTestData(mockData) {